Crate bp_relayers

Primitives of messages module.

Structs

ExtensionCallData

Extra post-dispatch data, associated with the supported runtime call.

PayRewardFromAccount

Reward payment procedure that executes a balances::transfer call from the account derived from the given RewardsAccountParams to the relayer or an alternative beneficiary.

Registration

Relayer registration.

RelayerRewardsKeyProvider

Can be used to access the runtime storage key within the RelayerRewards map of the relayers pallet.

RewardsAccountParams

Structure used to identify the account that pays a reward to the relayer.

RuntimeWithUtilityPallet

An BatchCallUnpacker adapter for runtimes with utility pallet.

Enums

ExplicitOrAccountParams

Either explicit account reference or RewardsAccountParams.

ExtensionCallInfo

Type of the call that the signed extension recognizes.

RewardsAccountOwner

The owner of the sovereign account that should pay the rewards.

Traits

BatchCallUnpacker

Something that can unpack batch calls (all-or-nothing flavor) of given size.

ExtensionConfig

Signed extension configuration.

PaymentProcedure

Reward payment procedure.

RewardLedger

A trait defining a reward ledger, which tracks rewards that can be later claimed.

StakeAndSlash

Relayer stake-and-slash mechanism.