Trait bp_relayers::ExtensionConfig

pub trait ExtensionConfig {
    type IdProvider: StaticStrProvider;
    type Runtime: Config;
    type BridgeMessagesPalletInstance: 'static;
    type PriorityBoostPerMessage: Get<TransactionPriority>;
    type Reward;
    type RemoteGrandpaChainBlockNumber: Clone + Copy + Debug;

    // Required methods
    fn parse_and_check_for_obsolete_call(
        call: &<Self::Runtime as SystemConfig>::RuntimeCall,
    ) -> Result<Option<ExtensionCallInfo<Self::RemoteGrandpaChainBlockNumber>>, TransactionValidityError>;
    fn check_obsolete_parsed_call(
        call: &<Self::Runtime as SystemConfig>::RuntimeCall,
    ) -> Result<&<Self::Runtime as SystemConfig>::RuntimeCall, TransactionValidityError>;
    fn check_call_result(
        call_info: &ExtensionCallInfo<Self::RemoteGrandpaChainBlockNumber>,
        call_data: &mut ExtensionCallData,
        relayer: &<Self::Runtime as SystemConfig>::AccountId,
    ) -> bool;
}

Signed extension configuration.

The single pallet-bridge-relayers instance may be shared by multiple messages pallet instances, bridging with different remote networks. We expect every instance of the messages pallet to add a separate signed extension to runtime. So it must have a separate configuration.

Required Associated Types

type IdProvider: StaticStrProvider

Unique identifier of the signed extension that will use this configuration.

type Runtime: Config

Runtime that optionally supports batched calls. We assume that batched call succeeds if and only if all of its nested calls succeed.

type BridgeMessagesPalletInstance: 'static

Messages pallet instance.

type PriorityBoostPerMessage: Get<TransactionPriority>

Additional priority that is added to base message delivery transaction priority for every additional bundled message.

type Reward

Type of reward, that the pallet-bridge-relayers is using.

type RemoteGrandpaChainBlockNumber: Clone + Copy + Debug

Block number for the remote GRANDPA chain. Mind that this chain is not necessarily the chain that we are bridging with. If we are bridging with parachain, it must be its parent relay chain. If we are bridging with the GRANDPA chain, it must be it.

Required Methods

fn parse_and_check_for_obsolete_call( call: &<Self::Runtime as SystemConfig>::RuntimeCall, ) -> Result<Option<ExtensionCallInfo<Self::RemoteGrandpaChainBlockNumber>>, TransactionValidityError>

Given runtime call, check if it is supported by the signed extension. Additionally, check if call (or any of batched calls) are obsolete.

fn check_obsolete_parsed_call( call: &<Self::Runtime as SystemConfig>::RuntimeCall, ) -> Result<&<Self::Runtime as SystemConfig>::RuntimeCall, TransactionValidityError>

Check if runtime call is already obsolete.

fn check_call_result( call_info: &ExtensionCallInfo<Self::RemoteGrandpaChainBlockNumber>, call_data: &mut ExtensionCallData, relayer: &<Self::Runtime as SystemConfig>::AccountId, ) -> bool

Given runtime call info, check that this call has been successful and has updated runtime storage accordingly.

Object Safety

This trait is not object safe.